# api-spec/src/main/resources/openapi/gox.yaml
openapi: 3.0.0
info:
  title: GoX API
  version: 1.0.0
  description: API for managing cars in the GoX car rental system.
servers:
  - url: http://localhost:8080/api
paths:
  /cars:
    get:
      summary: Retrieve a list of cars, with optional filtering
      operationId: getCars
      tags:
        - Car
      parameters:
        - in: query
          name: filter
          style: deepObject
          explode: true
          schema:
            $ref: '#/components/schemas/CarFilterRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Car'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new car
      operationId: createCar
      tags:
        - Car
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
...
  /users:
    get:
      summary: Retrieve all users
      operationId: getUsers
      tags:
        - User
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
  /users/{userId}:
    get:
      summary: Get a user by ID
      operationId: getUserById
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '404':
          description: Not Found
        '403':
          description: Forbidden
    put:
      summary: Update a user by ID
      operationId: updateUserById
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAdminUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a user by ID
      operationId: deleteUser
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /users/{userId}/bookings:
    get:
      summary: Get all bookings for a given user
      operationId: getBookingsByUserId
      tags:
        - Booking
      parameters:
        - name: userId
          in: path
          description: ID of the user whose bookings to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingSummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /users/{userId}/reviews:
    get:
      summary: Get all reviews by a given user
      operationId: getReviewsByUserId
      tags:
        - Review
      parameters:
        - name: userId
          in: path
          description: ID of the user whose reviews to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /users/me:
    get:
      summary: Get own profile
      operationId: getProfile
      tags:
        - User
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
    put:
      summary: Update own profile
      operationId: updateProfile
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
  /users/me/wishlist:
    get:
      summary: Get all car-IDs in the current user's Wishlist
      operationId: getWishlistCarIds
      tags:
        - User
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
  /users/me/wishlist/{carId}:
    post:
      summary: Add a car to the current user's Wishlist
      operationId: addCarToWishlist
      tags:
        - User
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Remove a car from the current user's Wishlist
      operationId: removeCarFromWishlist
      tags:
        - User
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /users/me/reviews:
    get:
      summary: Get all reviews by the current user
      operationId: getOwnReviews
      tags:
        - Review
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
  /users/me/bookings:
    get:
      summary: Get all bookings for current user
      operationId: getOwnBookings
      tags:
        - User
        - Booking
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingSummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
...
components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the car.
        brand:
          type: string
          description: Brand of the car.
        model:
          type: string
          description: Model of the car.
        year:
          type: integer
          description: Production year of the car.
        seats:
          type: integer
          description: Number of seats in the car.
        pricePerDay:
          type: number
          format: double
          description: Rental price per day.
        state:
          $ref: '#/components/schemas/CarState'
        category:
          $ref: '#/components/schemas/CarCategory'
        transmission:
          $ref: '#/components/schemas/TransmissionType'
        fuelType:
          $ref: '#/components/schemas/FuelType'
        description:
          type: string
          description: Optional description of the car.
        preview:
          $ref: '#/components/schemas/Photo'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        location:
          $ref: '#/components/schemas/Location'
    CarState:
      type: string
      description: The current state of the car.
      enum:
        - AVAILABLE
        - RENTED
        - UNAVAILABLE
        - IN_TRANSIT
    CarCategory:
      type: string
      enum:
        - SEDAN
        - SUV
        - HATCHBACK
        - COUPE
        - WAGON
        - CROSSOVER
        - SPORT
        - CABRIOLET
        - MUV
    CarFilterRequest:
      type: object
      properties:
        brand:
          type: string
        year:
          type: integer
          format: int32
        priceMin:
          type: number
          format: double
        priceMax:
          type: number
          format: double
        category:
          $ref: '#/components/schemas/CarCategory'
        seats:
          type: integer
          format: int32
        transmission:
          $ref: '#/components/schemas/TransmissionType'
        fuelType:
          $ref: '#/components/schemas/FuelType'
        sortBy:
         type: string
         description: one of brand, model, year, pricePerDay
         default: brand
        sortDir:
         type: string
         description: asc or desc
         default: asc
    TransmissionType:
      type: string
      enum:
        - MANUAL
        - AUTOMATIC
    FuelType:
      type: string
      enum:
        - PETROL
        - DIESEL
        - ELECTRIC
        - HYBRID
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the user.
        username:
          type: string
          description: Login user name
        name:
          type: string
          description: Full name of the user.
        email:
          type: string
          format: email
          description: Unique email of the user.
        role:
          $ref: '#/components/schemas/UserRole'
        loyaltyLevel:
          $ref: '#/components/schemas/LoyaltyLevel'
        address:
          type: string
          description: Postal address of the user.
        phone:
          type: string
          description: Phone number of the user.
        wishlist:
          $ref: '#/components/schemas/Wishlist'
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        loyaltyLevel:
          $ref: '#/components/schemas/LoyaltyLevel'
        address:
          type: string
        phone:
          type: string
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - CUSTOMER
      description: Role of the user within the system.
    UserUpdateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Full name of the user.
        phone:
          type: string
          description: Phone number of the user.
        address:
          type: string
          description: Postal address of the user.
    UserAdminUpdateRequest:
      type: object
      required:
        - name
        - phone
        - address
        - loyaltyLevel
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        loyaltyLevel:
          type: string
          description: One of STANDARD, SILVER, GOLD
    LoyaltyLevel:
      type: string
      enum:
        - STANDARD
        - SILVER
        - GOLD
      description: Loyalty level of the user.
    Wishlist:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique Wishlist ID.
        cars:
          type: array
          items:
            $ref: '#/components/schemas/Car'
      description: User's Wishlist with selected cars.
    Review:
      type: object
      properties:
        id:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        userId:
          type: integer
        userName:
          type: string
          description: Name of the user who wrote the review.
        carId:
          type: integer
        carBrand:
          type: string
        carModel:
          type: string
        createdAt:
          type: string
          format: date-time
          description: Date and time when the review was created.
        updatedAt:
          type: string
          format: date-time
          description: Date and time when the review was updated.
...
    Photo:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        isPreview: { type: boolean }
        url:
          type: string
          description: URL to download raw content of this photo
...
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        street:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    BookingSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        actualReturnDate:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/BookingStatus'
        carId:
          type: integer
          format: int64
        carBrand:
          type: string
        carModel:
          type: string
        userId:
          type: integer
          format: int64
        userName:
          type: string
...
    BookingStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - CANCELLED
        - COMPLETED
      description: Status of a Booking
...
    CarCreateRequest:
      type: object
      required:
        - brand
        - model
        - year
        - seats
        - pricePerDay
        - state
        - category
        - transmission
        - fuelType
        - locationId
      properties:
        brand:
          type: string
          description: Brand of the car.
        model:
          type: string
          description: Model of the car.
        year:
          type: integer
          format: int32
          description: Year of manufacture.
        seats:
          type: integer
          format: int32
          description: Number of seats in the car.
        pricePerDay:
          type: number
          format: double
          description: Daily rental price.
        state:
          $ref: '#/components/schemas/CarState'
        category:
          $ref: '#/components/schemas/CarCategory'
        transmission:
          $ref: '#/components/schemas/TransmissionType'
        fuelType:
          $ref: '#/components/schemas/FuelType'
        description:
          type: string
          description: Optional description of the car.
        locationId:
          type: integer
          format: int64
          description: ID of the car's current location.
...